requirements.txt
tweepy==4.12.1
oauthlib==3.2.2
requests==2.28.2
requests-oauthlib==1.3.1
charset-normalizer==3.0.1
idna==3.4
urllib3==1.26.14
certifi==2022.12.7

pytz==2022.7.1

python-dotenv==0.21.1

flake8==6.0.0
mccabe==0.7.0
pycodestyle==2.10.0
pyflakes==3.0.1
isort==5.12.0